+++ GUI/MainWindow.xaml
@@ -4,7 +4,6 @@
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
         xmlns:local="clr-namespace:GUI"
-        
         mc:Ignorable="d"
         d:DataContext="{Binding Source={d:DesignInstance Type=local:MainViewModel}}"
         Title="MainWindow" Height="1080" Width="1920">
